Update patch set 1
Patch Set 1:

When you specify that the deployment is using generated certificates, then it will override those password properties and use the .password and .passphrases file. 
We need to keep this until we are completely migrated to using aaf generated passwords.
Its not obsolete because that isn't first delivered first of all and we need to be flexible so if that solution doesn't work out, we can switch easily.

Patch-set: 1
Reviewer: Gerrit User 75 <75@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0